fix: address code review - rating precision and magic numbers
Agent-Logs-Url: https://github.com/merenyimiklos/fridgeIQ/sessions/64367095-b902-459e-b893-92a00252e2ee

+++ lib/features/meal_suggestion/presentation/widgets/recipe_match_card.dart
@@ -101,7 +101,7 @@ class RecipeMatchCard extends ConsumerWidget {
               data: (reviews) {
                 if (reviews.isEmpty) return const SizedBox.shrink();
                 final avgRating =
-                    reviews.fold<int>(0, (sum, r) => sum + r.rating) /
+                    reviews.fold<double>(0.0, (sum, r) => sum + r.rating) /
                         reviews.length;
                 return Padding(
                   padding: const EdgeInsets.only(top: 4),
